requirements.txt
django-allauth==0.44.0
Django~=3.1.5
django-allauth~=0.44.0
yahoo-fin==0.8.6
pandas==1.2.0
lxml==4.6.2
-e git://github.com/Fantomas42/django-blog-zinnia.git@881101a9d1d455b2fc581d6f4ae0947cdd8126c6#egg=django_blog_zinnia
requests-html==0.10.0
django-contrib-comments~=2.0.0
django-tagging~=0.5.0